requirements.txt
httpx[http2]
pyexecjs
lxml
fake_useragent